remove quantinfo, move new checkpoint to check_point_v2.py
