+++ oneflow/core/framework/variable_meta_info.proto
@@ -4,12 +4,7 @@ package oneflow;
 import "oneflow/core/common/shape.proto";
 import "oneflow/core/common/data_type.proto";
 
-message QuantInfo {
-  
-}
-
 message VariableMetaInfo {
   required ShapeProto shape = 2;
   required DataType data_type = 3;
-  optional QuantInfo quant_info = 4;
 }
